Use more extern templates; create core/equal
