--- compiler+runtime/include/cpp/jank/prelude.hpp
@@ -1,40 +1,0 @@
-#pragma once
-
-#include <random>
-#include <chrono>
-#include <unordered_map>
-
-#include <gc/gc.h>
-#include <gc/gc_cpp.h>
-#include <gc/gc_allocator.h>
-
-#include <immer/heap/gc_heap.hpp>
-#include <immer/heap/heap_policy.hpp>
-#include <immer/memory_policy.hpp>
-
-#include <nanobench.h>
-
-#include <jank/type.hpp>
-#include <jank/profile/time.hpp>
-#include <jank/util/make_array.hpp>
-#include <jank/util/scope_exit.hpp>
-#include <jank/runtime/object.hpp>
-#include <jank/runtime/detail/type.hpp>
-#include <jank/runtime/rtti.hpp>
-#include <jank/runtime/core.hpp>
-#include <jank/runtime/context.hpp>
-#include <jank/runtime/behavior/seqable.hpp>
-#include <jank/runtime/behavior/collection_like.hpp>
-#include <jank/runtime/behavior/sequential.hpp>
-#include <jank/runtime/behavior/number_like.hpp>
-#include <jank/runtime/behavior/nameable.hpp>
-#include <jank/runtime/behavior/derefable.hpp>
-#include <jank/runtime/behavior/transientable.hpp>
-#include <jank/runtime/behavior/conjable.hpp>
-#include <jank/runtime/behavior/associatively_writable.hpp>
-#include <jank/runtime/behavior/stackable.hpp>
-#include <jank/runtime/behavior/indexable.hpp>
-#include <jank/runtime/behavior/chunkable.hpp>
-#include <jank/runtime/behavior/comparable.hpp>
-
-#include <jank/c_api.h>
