+++ compiler+runtime/src/cpp/clojure/core_native.cpp
@@ -4,6 +4,7 @@
 #include <clojure/core_native.hpp>
 #include <jank/runtime/convert.hpp>
 #include <jank/runtime/core.hpp>
+#include <jank/runtime/core/equal.hpp>
 #include <jank/runtime/core/meta.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/behavior/callable.hpp>
