+++ compiler+runtime/src/cpp/jank/error.cpp
@@ -2,7 +2,7 @@
 
 #include <jank/error.hpp>
 #include <jank/runtime/context.hpp>
-#include <jank/runtime/core.hpp>
+#include <jank/runtime/core/to_string.hpp>
 #include <jank/runtime/obj/keyword.hpp>
 
 namespace jank::error
