+++ compiler+runtime/src/cpp/jank/runtime/detail/native_persistent_array_map.cpp
@@ -1,7 +1,7 @@
 #include <fmt/format.h>
 
 #include <jank/runtime/detail/native_persistent_array_map.hpp>
-#include <jank/runtime/core.hpp>
+#include <jank/runtime/core/equal.hpp>
 
 namespace jank::runtime::detail
 {
