+++ compiler+runtime/src/cpp/jank/runtime/object.cpp
@@ -1,5 +1,5 @@
 #include <jank/runtime/object.hpp>
-#include <jank/runtime/core.hpp>
+#include <jank/runtime/core/equal.hpp>
 #include <jank/runtime/visit.hpp>
 #include <jank/hash.hpp>
