+++ compiler+runtime/test/cpp/jank/jit/processor.cpp
@@ -16,6 +16,7 @@
 #include <jank/runtime/obj/keyword.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/core/to_string.hpp>
+#include <jank/runtime/core/equal.hpp>
 #include <jank/analyze/processor.hpp>
 #include <jank/jit/processor.hpp>
