+++ compiler+runtime/test/cpp/jank/runtime/behavior/callable.cpp
@@ -1,6 +1,6 @@
 #include <jank/runtime/behavior/callable.hpp>
-#include <jank/runtime/core.hpp>
 #include <jank/runtime/core/make_box.hpp>
+#include <jank/runtime/core/equal.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/obj/persistent_list.hpp>
 #include <jank/runtime/obj/persistent_string.hpp>
