+++ compiler+runtime/test/cpp/jank/runtime/core/equal.cpp
@@ -1,5 +1,5 @@
-#include <jank/runtime/core.hpp>
 #include <jank/runtime/core/make_box.hpp>
+#include <jank/runtime/core/equal.hpp>
 
 /* This must go last; doctest and glog both define CHECK and family. */
 #include <doctest/doctest.h>
